requirements.txt
streamlit
openai
python-dotenv
streamlit-webrtc
numpy
av
